refactor: divide into packages
