+++ src/main/java/baseball/result/ResultType.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.result;
 
 public enum ResultType {
     EXACT_MATCH, BALL_AND_STRIKE, ONLY_BALL, ONLY_STRIKE, NOTHING
