+++ src/main/java/baseball/service/Validator.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.service;
 
 import java.util.Arrays;
 import java.util.HashSet;
